requirements.txt
aiohttp>=3.8.1
matplotlib>=3.5.1
numpy>=1.21.0
pandas>=1.3.0
requests>=2.25.1
PyYAML>=6.0
PyQt5>=5.15.2
beautifulsoup4>=4.10.0
python-telegram-bot>=20.0
